% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzightmap_modifiers.R
\name{addLayer.iNZightMapPlot}
\alias{addLayer.iNZightMapPlot}
\title{Add a ggplot layer to an iNZightMapPlot object}
\usage{
addLayer.iNZightMapPlot(mapplot.obj, layer.set, layer.name, layer.obj)
}
\arguments{
\item{mapplot.obj}{An iNZightMapPlot object.}

\item{layer.set}{Which set of layers to add to --- \code{"map"} or \code{"region"}.}

\item{layer.name}{Name to give to the added layer.}

\item{layer.obj}{The layer object to add.}
}
\value{
An iNZightMapPlot object with \code{layer.obj} stored within it.
}
\description{
Add a ggplot layer to an iNZightMapPlot object
}
\examples{

}
