% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{bar.coor}
\alias{bar.coor}
\title{create bar object}
\usage{
bar.coor(obj, var, data, xmax = 0.85, ymax = 2,
  bar.col = c("#E0FFFF", "#FAFAD2", "#FFA07A", "#C71585", "#DC143C",
  "#B8860B", "#00BFFF", "#ADFF2F"))
}
\arguments{
\item{obj}{a inzightshapemap object}

\item{var}{a character vectror, the column name of data}

\item{data}{the data set}

\item{xmax}{the maximum weight of the bar}

\item{ymax}{the maximum height of the bar}

\item{bar.col}{the color of each bar}
}
\description{
a function for create a bar object
}
\author{
Jason
}
