% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ClickOnZoom}
\alias{ClickOnZoom}
\title{Zoom in/out}
\usage{
ClickOnZoom(...)
}
\arguments{
\item{...}{additional parameters}
}
\description{
Zoom in/out
}
\details{
if ratio < 1 then zoom in, if ratio > 1 then zoom out, if ratio = 1 then shift the plot.
}
\author{
Jason
}
