% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{col.fun}
\alias{col.fun}
\title{Color Specification}
\usage{
col.fun(data, color.index, display = "hue", na.fill = "white", offset = 0,
  col = "red")
}
\arguments{
\item{data}{a numeric vector that should lie on the range of [0,1]}

\item{color.index}{an integer vector}

\item{display}{a character value, the method for display colors. It should be one value of "hcl","hue","heat","rainbow","terrain","topo","cm","gray","r","n"}

\item{na.fill}{a character value, fill the unmatch region/country by the color.}

\item{offset}{a numeric value within the range of [0,1]}

\item{col}{the color for fill the match region/country, it only needs to be specify if display = 'hue'}
}
\value{
an color vector
}
\description{
Choose a type of color that helps read the map nicely.
}
\author{
Jason Wen
}

