% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{color.bind}
\alias{color.bind}
\title{Plot an iNZight Map}
\usage{
color.bind(color, obj)
}
\arguments{
\item{color}{a numeric vector see\link{rgb}}

\item{obj}{an shape obj}
}
\value{
a list that contain shape object and colors for filling
}
\description{
bind the color and a shape object together
}
\details{
the function will also returns a global object which called global.objects
}
\author{
Jason Wen
}

