% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzmap.R
\name{create.inz.mapplot}
\alias{create.inz.mapplot}
\title{title}
\usage{
\method{create}{inz.mapplot}(obj, ...)
}
\arguments{
\item{obj}{an object from within iNZightPlot}

\item{...}{additional arguments (ignored)}
}
\value{
Object
}
\description{
title
}
\author{
Tom Elliott
}
