% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzshapemap.R
\name{create.inz.shapemapplot}
\alias{create.inz.shapemapplot}
\title{draw a shape file ...}
\usage{
\method{create}{inz.shapemapplot}(obj)
}
\arguments{
\item{obj}{an object from within iNZightPlot}
}
\value{
Object
}
\description{
details...
}
\author{
Tom Elliott
}

