% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{data.trans}
\alias{data.trans}
\title{Transform the data into the range of [0,1].}
\usage{
data.trans(x, transform = "linear", data.range, mean, sd, max.prob)
}
\arguments{
\item{x}{a numeric value or vector.}

\item{transform}{the method for transformation, can be linear,log,sqrt,exp,power and normal.}

\item{data.range}{??? range of \code{x}}

\item{mean}{??? mean of \code{x}}

\item{sd}{??? standard deviation of \code{x}}

\item{max.prob}{???}
}
\value{
a transformed numeric vector.
}
\description{
Transform the data into the range of [0,1].
}
\author{
Jason Wen
}
