% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{decodeMapDir}
\alias{decodeMapDir}
\title{Convert map filepaths to "tidy" filepaths}
\usage{
decodeMapDir(mapdir.mat)
}
\arguments{
\item{mapdir.mat}{Matrix of metadata}
}
\description{
Convert map filepaths to "tidy" filepaths
}
