% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMapGG.R
\name{dot.density}
\alias{dot.density}
\title{Generate dots randomly within regions}
\usage{
dot.density(data, var, total = 300, per.dot = 1000, prop = FALSE)
}
\arguments{
\item{data}{Map dataset}

\item{var}{Variable to use to allocate dots}

\item{total}{For proportion dot density maps, the total number of points}

\item{per.dot}{For "dot per N" maps, the N that each dot reprsents}

\item{prop}{Boolean to switch between proportional dot density/"dot per N" maps.}
}
\value{
An sf object containing a series of dots for each region.
}
\description{
Generate dots randomly within regions
}
