% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{download.shapefiles}
\alias{download.shapefiles}
\title{Download shapefiles from remote repository}
\usage{
download.shapefiles(dirURL, currPath, shapefileDir)
}
\arguments{
\item{dirURL}{URL of the directory containing shapefiles}

\item{currPath}{Current path to save to}
}
\description{
Download shapefiles from remote repository
}
