% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{drawing.features}
\alias{drawing.features}
\title{drawing features}
\usage{
drawing.features(bar.obj, latlon, cols, shade.each, region.name, data.region,
  value, name, center.x, center.y, y.shift = 0.5)
}
\arguments{
\item{bar.obj}{a bar object see \link{bar.coor}}

\item{latlon}{a n*2 numeric matrix the first column specifies the latitudes and the second column specifies the longitudes}

\item{cols}{a color character strings vector}

\item{shade.each}{a numeric vector}

\item{region.name}{a character vector}

\item{data.region}{region data}

\item{value}{a numeric vector}

\item{name}{a character vector}

\item{center.x}{a numeric value}

\item{center.y}{a numeric value}

\item{y.shift}{a numeric value}
}
\description{
a function for drawing bar char, display the value and/or region name
}
\author{
Jason
}
