% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{findBestMatch}
\alias{findBestMatch}
\title{Find 'best-matching' pair of variables from map and data}
\usage{
findBestMatch(data, map.data)
}
\arguments{
\item{data}{Data frame of the input data}

\item{map.data}{sf object of the map data}
}
\value{
Two-element vector of form (data variable, map variable)
}
\description{
Find 'best-matching' pair of variables from map and data
}
