% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparklines.R
\name{geom_sparkline}
\alias{geom_sparkline}
\title{Sparkline layer for each map point}
\usage{
geom_sparkline(
  mapping = ggplot2::aes(),
  data = NULL,
  stat = "sf",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them.}

\item{...}{Other arguments passed on to layer().}
}
\value{
A sparkline layer
}
\description{
Plots a minimalist time series line chart at each coordinate
provided
}
\details{
`geom_sparkline` plots a small time series chart on points on an `sf` map.
}
