% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{Get.map.size}
\alias{Get.map.size}
\title{title}
\usage{
Get.map.size(latR.odd, lonR.odd, SCALE)
}
\arguments{
\item{latR.odd}{}

\item{lonR.odd}{}

\item{SCALE}{}
}
\value{
something
}
\description{
desc
}
\details{
det
}
\author{
tell029
}

