% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getMinMax}
\alias{getMinMax}
\title{Get the range of a set of variables while ignoring categorical variables}
\usage{
getMinMax(obj, vars)
}
\arguments{
\item{obj}{iNZightMapPlot object}

\item{vars}{Vector of variables to calculate range of}
}
\value{
Vector of ranges for each variable
}
\description{
Get the range of a set of variables while ignoring categorical variables
}
