% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{getNewMap}
\alias{getNewMap}
\title{title}
\usage{
getNewMap(xlim, ylim, SCALE, type, zoom)
}
\arguments{
\item{xlim}{}

\item{ylim}{}

\item{SCALE}{}

\item{type}{}

\item{zoom}{}
}
\value{
something
}
\description{
desc
}
\details{
det
}
\author{
tell029
}

