% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{innerLim}
\alias{innerLim}
\title{inner limit}
\usage{
innerLim(obj, d.region)
}
\arguments{
\item{obj}{an inzshapemap object.}

\item{d.region}{a character vector that specify the region or country.}
}
\value{
a numeric vector with length 4, the inner limit/bbox of the map.
}
\description{
Calculate the inner limit.
}
\details{
the limit is computed by calculate the limit of the given region.
}
\examples{
los = 'C:/Users/yeamin/Documents/GitHub/iNZightMaps/data/world.rds'
lod = 'C:/Users/yeamin/Desktop/Gapminder-2008.csv'
data = read.csv(lod,skip = 1)
obj = iNZightShapeMap(los, data.region = 'Country', data = data)
d.region = data$Country
innerLim(obj,d.region)
}
\author{
Jason Wen
}

