% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapobject.R
\name{iNZightMap}
\alias{iNZightMap}
\alias{plot.inzightmap}
\title{Create an iNZight Map Object}
\usage{
iNZightMap(lat, lon, data, name = deparse(substitute(data)))

\method{plot}{inzightmap}(
  x,
  opacity,
  map.zoom = -1,
  type = ggmap:::STADIA_VALID_MAP_TYPES,
  ...,
  env = parent.frame()
)
}
\arguments{
\item{lat}{a formula specifying the name of the latitude variable in \code{data}}

\item{lon}{a formula specifying the name of the longitude variable in \code{data}}

\item{data}{a data frame}

\item{name}{the name of the data set}

\item{x}{an \code{inzightmap} object}

\item{opacity}{character or expression of the variable name to code point opacity by}

\item{map.zoom}{the level of zoom to use; this should depend on the scales unless specified}

\item{type}{the type of map to download from Google}

\item{...}{additional arguments passed to \code{iNZightPlot}}

\item{env}{Environment to evaluate \code{iNZightPlot} in}
}
\value{
data frame object with class \code{inzightmap}
}
\description{
Create an iNZight Map Object and pass to the plot
}
\details{
The plot will download a map that contains all of the points. The size is limited by
the fixed zoom levels provided by Google's Static Maps API.
}
\section{Functions}{
\itemize{
\item \code{plot(inzightmap)}: Plot an \code{inzightmap} object

}}
\author{
Tom Elliott
}
