% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMapGG.R
\name{iNZightMapAggregation}
\alias{iNZightMapAggregation}
\title{Aggregate an iNZightMapPlot with Multiple Observations}
\usage{
iNZightMapAggregation(obj, aggregation = "mean", single.value = NULL)
}
\arguments{
\item{obj}{iNZightMapPlot object}

\item{aggregation}{Type of aggregation to use}

\item{single.value}{If \code{aggregation = "singlevalue"}, the observation that should be extracted}
}
