% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iNZightMapCountryISO}
\alias{iNZightMapCountryISO}
\title{Get list of country ISO codes}
\usage{
iNZightMapCountryISO()
}
\value{
Data frame with country names and ISO codes
}
\description{
Get list of country ISO codes
}
