% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMapGG.R
\name{iNZightMapPlot}
\alias{iNZightMapPlot}
\alias{iNZightMapPlotRegion}
\title{Create iNZightMapPlot object}
\usage{
iNZightMapPlot(data, map, type, ...)

iNZightMapPlotRegion(data, map, by.data, by.map, simplification.level = 0.01,
  multiple.obs = FALSE, sequence.var = NULL, agg.type = "last")
}
\description{
Create iNZightMapPlot object
}
\section{Functions}{
\itemize{
\item \code{iNZightMapPlotRegion}: Constructs a iNZightMapPlot using region values.
}}

