% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iNZightMapProjections}
\alias{iNZightMapProjections}
\title{Get available map projections}
\usage{
iNZightMapProjections()
}
\value{
Data frame of map projections
}
\description{
Get available map projections
}
