% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMapGG.R
\name{iNZightMapRegions}
\alias{iNZightMapRegions}
\title{Extract region names from iNZightMapPlot object}
\usage{
iNZightMapRegions(obj)
}
\arguments{
\item{obj}{iNZightMapPlot object}
}
\value{
Sorted vector of the region names
}
\description{
Extract region names from iNZightMapPlot object
}
