% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMapGG.R
\name{iNZightMapVars}
\alias{iNZightMapVars}
\title{Extract column names from an iNZightMapPlot object}
\usage{
iNZightMapVars(obj, map.vars = FALSE)
}
\arguments{
\item{obj}{iNZightMapPlot object}

\item{map.vars}{Should the variables included in the original map be included in the output?}
}
\description{
Extract column names from an iNZightMapPlot object
}
