% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapemapobject.R
\name{iNZightShapeMap}
\alias{iNZightShapeMap}
\title{Create an iNZight Shape Map Object}
\usage{
iNZightShapeMap(shp.RDS, location, column.index, region, data)
}
\arguments{
\item{location}{the location}
}
\description{
description ...
}
\details{
details ....
}
\examples{

}

