% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{is.google.map}
\alias{is.google.map}
\title{Is Google Map?}
\usage{
is.google.map(lat, lon)
}
\arguments{
\item{lat}{a numeric vector}

\item{lon}{a numeric vector}
}
\value{
return an logical value that tells wheater the coordinates are all within the same range of google map's or not.
}
\description{
The coordinates should be within on the same range of Google Map's. i.e. longitude = [-180,180], latitude = [-90,90]
}
\author{
Jason Wen
}
