% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{latlon.xy}
\alias{latlon.xy}
\title{latlon.xy}
\usage{
latlon.xy(data, map)
}
\arguments{
\item{data}{the data set that use for plotting, the first cloumn needs to be Longitude and the second column needs to be Latitude.}

\item{map}{a map object from RgoogleMaps package.}
}
\value{
a list that contain the tramsformed x and y.
}
\description{
Transform the latitude and longitude into xy
}
\details{
wrap function for \link{LatLon2XY} from 'RgoogleMaps' package
}
\author{
Jason Wen
}

