% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{lon.rescale}
\alias{lon.rescale}
\title{Transform the longitude}
\usage{
lon.rescale(lon)
}
\arguments{
\item{lon}{a numeric value or a numeric vector of longitude}
}
\value{
a new longitude value or vector that used for plotting in map.
}
\description{
Transform longitude in order to get the biggest map as possible
}
\author{
Jason Wen
}
