% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matchVariables}
\alias{matchVariables}
\title{Match two vectors}
\usage{
matchVariables(data.vect, map.vect)
}
\arguments{
\item{data.vect}{Vector containing the dataset variable}

\item{map.vect}{Vector containing the map object variable}
}
\value{
A list containing: the dataset variable vector with duplicates removed, which values in the dataset and map variable had a match, how many matches occured overall and if there was multiple observations for any region.
}
\description{
Match two vectors
}
