% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{name.match}
\alias{name.match}
\title{Match names between map and data}
\usage{
name.match(shp.region, data.region)
}
\arguments{
\item{shp.region}{Map object}

\item{data.region}{Data}
}
\value{
a 2*2 numeric matrix
}
\description{
Match names between map and data
}
\author{
Jason
}
