% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NSSATS}
\alias{NSSATS}
\title{National Survey of Substance Abuse Treatment Services}
\format{A data frame with 57 rows and 25 variables:
\describe{
  \item{STATE}{Two character state code}
  \item{OfferNonEnglish}{Proportion of facilities offering non-English language treatment}
  \item{SpecificGroupAdolescent}{Proportion of facilities offering specific groups for adolescents}
  \item{SpecificGroupWomen}{Proportion of facilities offering specific groups for women}
  \item{SpecificGroupPregnant}{Proportion of facilities offering specific groups for pregnant women}
  \item{SpecificGroupMen}{Proportion of facilities offering specific groups for men}
  \item{SpecificGroupSeniors}{Proportion of facilities offering specific groups for seniors}
  \item{SpecificGroupLGBT}{Proportion of facilities offering specific groups for LGBT}
  \item{SpecificGroupVeterans}{Proportion of facilities offering specific groups for veterans}
  \item{SpecificGroupMentalIllness}{Proportion of facilities offering specific groups for individuals with mental illness}
  \item{SpecificGroupTrauma}{Proportion of facilities offering specific groups for individuals with trauma}
  \item{SlidingFeeScale}{Proportion of facilities offering a sliding fee scale}
  \item{OffersFreeTreatment}{Proportion of facilities offering free treatment to those who cannot pay}
  \item{AcceptsMedicare}{Proportion of facilities that accept Medicare as a payment option}
  \item{AcceptsMedicaid}{Proportion of facilities that accept Medicaid as a payment option}
  \item{AcceptsStateInsurance}{Proportion of facilities that accept state insurance as a payment option}
  \item{AcceptsMilitaryInsurance}{Proportion of facilities that accept military insurance as a payment option}
  \item{AcceptsPrivateInsurance}{Proportion of facilities that accept private insurance as a payment option}
  \item{LicensedByAuthority}{Proportion of facilities licensed/certified/accredited by hospital or state authority}
  \item{PrivateForProfit}{Proportion of facilities that are private for-profit organisations}
  \item{PrivateNonProfit}{Proportion of facilities that are private non-profit organisations}
  \item{StateGovt}{Proportion of facilities owned by the state government}
  \item{DUIProgram}{Proportion of facilities offering a program for DUI/DWI offenders}
  \item{OfferSignLanguage}{Proportion of facilities offering treatment in sign language}
  \item{AnyPublicFunding}{Proportion of facilities that receive any federal, state, county, or local funds}
}}
\source{
\href{https://wwwdasis.samhsa.gov/dasis2/nssats.htm}{National Survey of Substance Abuse Treatment Services (N-SSATS)}
}
\usage{
NSSATS
}
\description{
A dataset containing the proportion of substance abuse treatment facilities in 
a given state that meet a criteria given by the variable name.
}
\keyword{datasets}
