% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzquakes.R
\docType{data}
\name{nzquakes}
\alias{nzquakes}
\title{Earthquakes in New Zealand}
\format{A data frame with 334 rows and 12 variables:
\describe{
  \item{Order}{For ordering the observations}
  \item{Latitude}{The latitude ("Y") positions of the earthquakes}
  \item{Longitude}{The longitude ("X") positions of the earthquakes}
}}
\usage{
nzquakes
}
\description{
A dataset containing locations and measurements of earthquakes in the
New Zealand region.
}
\keyword{datasets}

