% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{order.match}
\alias{order.match}
\title{re-order the region name in shape file}
\usage{
order.match(shp.region, data.region)
}
\arguments{
\item{shp.region}{a character vector.}

\item{data.region}{a character vector.}
}
\value{
an integer vector.
}
\description{
re-order the region name in shape file
}
\examples{
los = 'C:/Users/yeamin/Documents/GitHub/iNZightMaps/data/world.rds'
shp = readRDS(los)
shp.region = shp$region
lod = 'C:/Users/yeamin/Desktop/Gapminder-2008.csv'
data = read.csv(lod,skip = 1)
order.match(shp.region,data$Country)
}
\author{
Jason Wen
}
