% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{outerLim}
\alias{outerLim}
\title{outer limit}
\usage{
outerLim(obj, lim, ignore.region = c("Russia", "Antarctica", "New Zealand"))
}
\arguments{
\item{obj}{an inzshapemap object.}

\item{lim}{a numeric vector of length 4.}

\item{ignore.region}{a character value or vector, sepecify which regions are been ignored when calculate the limit.}
}
\value{
a numeric vector with length 4, the outer limit/bbox of the map.
}
\description{
Calculate the outer limit.
}
\details{
the limit is computed by calculate the limit of the given inner limit. If there is a region been cut by the inner limit, then the outer limit will extend up to the limit of the region.
}
\author{
Jason Wen
}

