% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapobject.R
\name{plot.inzightmap}
\alias{plot.inzightmap}
\title{Plot iNZight Map Object}
\usage{
\method{plot}{inzightmap}(x, ...)
}
\arguments{
\item{x}{an object of class \code{inzightmap}}

\item{...}{extra arguments sent to \code{iNZightPlot}}
}
\value{
iNZightPlot object
}
\description{
plot iNZight Map Object
}
\details{
None.
}
\author{
Tom Elliott
}

