% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzightmap.R
\name{plot.iNZightMapPlot}
\alias{plot.iNZightMapPlot}
\title{Plot an iNZightMapPlot object}
\usage{
\method{plot}{iNZightMapPlot}(obj)
}
\arguments{
\item{obj}{An iNZightMapPlot object of any type.}
}
\description{
Plot an iNZightMapPlot object
}
\examples{

}
