% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapemapobject.R
\name{plot.inzightshapemap}
\alias{plot.inzightshapemap}
\title{Create an iNZight Shape Map Object}
\usage{
\method{plot}{inzightshapemap}(
  x,
  variable,
  col.fun = "hue",
  transform = "linear",
  col.offset = 0.2,
  col = "red",
  na.fill = "#F4A460",
  full.map = TRUE,
  extend.ratio = 1,
  name = FALSE,
  zoom = 1,
  zoom.center = c(NA, NA),
  ...
)
}
\arguments{
\item{x}{the iNZight Shape Map Object}

\item{variable}{the variable or the column name in the data}

\item{col.fun}{a character value corresponding to the colour scale function to use}

\item{transform}{Parameters for \code{inzpar}}

\item{col.offset}{Parameters for \code{inzpar}}

\item{col}{Parameters for \code{inzpar}}

\item{na.fill}{Fill colour of regions with NA for \code{variable}}

\item{full.map}{logical value.}

\item{extend.ratio}{Parameters for \code{inzpar}}

\item{name}{Parameters for \code{inzpar}}

\item{zoom}{Parameters for \code{inzpar}}

\item{zoom.center}{Parameters for \code{inzpar}}

\item{...}{Additional arguments}
}
\description{
Create an iNZightShapeMap object
}
\details{
details ....
}
\author{
Tom Elliott
}
