% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzmap.R
\name{plot.inzmap}
\alias{plot.inzmap}
\title{Plot an iNZight Map}
\usage{
\method{plot}{inzmap}(x, gen, ...)
}
\arguments{
\item{gen}{other options passed from iNZightPlot}

\item{...}{additional arguments (ignored)}

\item{obj}{object passed from iNZightPlot}
}
\description{
draw a map by passing an iNZightPlot object
}
\details{
the function will also returns a global object which called global.objects
}
\author{
Jason Wen
}
