% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzshapemap.R
\name{plot.inzshapemap}
\alias{plot.inzshapemap}
\title{Plot an iNZight Shape Map}
\usage{
\method{plot}{inzshapemap}(x, gen, ...)
}
\arguments{
\item{x}{object passed from iNZightPlot}

\item{gen}{other options passed from iNZightPlot}

\item{...}{additional arguments (ignored)}
}
\description{
draw a shape file ...
}
\details{
details...
}
\author{
Jason Wen
}
