% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{polygons_per_region}
\alias{polygons_per_region}
\title{Get number of polygons for each region of a map}
\usage{
polygons_per_region(obj)
}
\arguments{
\item{obj}{iNZightMapPlot object}
}
\value{
Vector containing the number of polygons for each region of the map
}
\description{
Get number of polygons for each region of a map
}
