% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{re.scale}
\alias{re.scale}
\title{rearrange the limit}
\usage{
re.scale(x, ratio)
}
\arguments{
\item{x}{a vector of length 2.}

\item{ratio}{a numeric value.}
}
\value{
a vector of length of 2, re-sized by the ratio.
}
\description{
rearrange the limit by given a ratio.
}
\author{
Jason Wen
}

