% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read.mapmetadata}
\alias{read.mapmetadata}
\title{Read shapefile metadata from file}
\usage{
read.mapmetadata(shapefileDir)
}
\arguments{
\item{shapefileDir}{Directory containing the shapefiles for iNZight}
}
\value{
Matrix of filepaths, filenames and descriptions of each shapefile available in `shapefileDir`
}
\description{
Read shapefile metadata from file
}
