% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzightmap_modifiers.R
\name{removeLayer.iNZightMapPlot}
\alias{removeLayer.iNZightMapPlot}
\title{Remove a ggplot layer from an iNZightMapPlot object}
\usage{
removeLayer.iNZightMapPlot(mapplot.obj, layer.set, layer.name)
}
\arguments{
\item{mapplot.obj}{An iNZightMapPlot object.}

\item{layer.set}{Which set of layers the layer is contained in --- \code{"map"} or \code{"region"}.}

\item{layer.name}{Name of the layer to be removed.}
}
\description{
Remove a ggplot layer from an iNZightMapPlot object
}
\examples{

}
