% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{retrieveMap}
\alias{retrieveMap}
\title{Retrieve a map for use in iNZightMapPlot}
\usage{
retrieveMap(filename)
}
\arguments{
\item{filename}{Filename of the map}
}
\value{
An sf object containing the map
}
