% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rezoom}
\alias{rezoom}
\title{change the zoom within the center point}
\usage{
rezoom(zoom)
}
\arguments{
\item{zoom}{a numeric value between 0.1 to 0.9(minimum zoom to maximum zoom)}
}
\description{
change the zoom within the center point
}
\author{
Jason
}

