% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzightmap_modifiers.R
\name{setMapping.iNZightMapPlot}
\alias{setMapping.iNZightMapPlot}
\title{Set an aesthetic mapping to a layer in an iNZightMapPlot object}
\usage{
setMapping.iNZightMapPlot(mapplot.obj, layer.set, layer.name = "baselayer",
  aes.name, aes.var)
}
\arguments{
\item{mapplot.obj}{An iNZightMapPlot object.}

\item{layer.set}{Which set of layers the layer is contained in ---
\code{"map"} or \code{"region"}.}

\item{layer.name}{Name of the layer to be modified. Must be a
\code{geom_sf()} layer.}

\item{aes.name}{The name of the aesthetic to set}

\item{aes.var}{The variable to map to the aesthetic}
}
\value{
An iNZightMapPlot object with the desired changes to
  \code{layer.name}.
}
\description{
This function modifies an aesthetic mapping to an already existing ggplot
layer in an iNZightMapPlot object. By replacing the existing layer within the
iNZightMapPlot object with a new layer that is identical aside from the new
mapping, multiple mapping adjustments will not inflate the iNZightMapPlot
object size nor leave unnecessary layers that would slow down plotting.
}
