% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{shape.extract}
\alias{shape.extract}
\title{Plot an iNZight Map}
\usage{
shape.extract(shp, column.index. = 2)
}
\arguments{
\item{shp}{a SpatialPolygonsDataFrame object, see \link{readShapeSpatial}}
}
\value{
a shape object
}
\description{
extract the information from a SpatialPolygonsDataFrame object and
}
\details{
the function will also returns a global object which called global.objects
}
\author{
Jason Wen
}

