% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{srezoom}
\alias{srezoom}
\title{change the zoom within the center point}
\usage{
srezoom(zoom)
}
\arguments{
\item{zoom}{a numeric value between 0.1 to 0.9(minimum zoom to maximum zoom)}
}
\description{
change the zoom within the center point
}
\author{
Jason
}

