% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{subByLim}
\alias{subByLim}
\title{subset by limit}
\usage{
subByLim(obj, lim)
}
\arguments{
\item{obj}{an inzshapemap object.}

\item{lim}{a numeric vector of length 4.}
}
\value{
an inzshapemap object.
}
\description{
Subset by limit.
}
\examples{
los = 'C:/Users/yeamin/Documents/GitHub/iNZightMaps/data/world.rds'
lod = 'C:/Users/yeamin/Desktop/Gapminder-2008.csv'
data = read.csv(lod,skip = 1)
obj = iNZightShapeMap(los, data.region = 'Country', data = data)
lim = c(-40,40,-40,40)
subByLim(obj,lim)
}
\author{
Jason Wen
}
