% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzightmap.R
\name{summary.iNZightMapPlot}
\alias{summary.iNZightMapPlot}
\title{Summarise an iNZightMapPlot object}
\usage{
\method{summary}{iNZightMapPlot}(obj)
}
\arguments{
\item{obj}{An iNZightMapPlot object of any type.}
}
\description{
Summarise an iNZightMapPlot object
}
\examples{

}
