% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{win.ratio}
\alias{win.ratio}
\title{win.ratio}
\usage{
win.ratio(xlim, ylim)
}
\arguments{
\item{xlim}{a numeric vector of length 2}

\item{ylim}{a numeric vector of length 2}
}
\value{
a numeric vector of length 4, the first two componets specified the new xlim and the last two componets specified the new ylim.
}
\description{
compute and return the xlim and ylim within aspect ratio
}
\examples{
xlim = c(-90,90)
ylim = c(-80,90)
grid.newpage()
win.ratio(xlim,ylim)
}
\author{
Jason
}

