% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ClickOnZoom}
\alias{ClickOnZoom}
\title{Zoom in/out}
\usage{
ClickOnZoom(ratio = 1, resize = FALSE,
  p.center = global.objects$maps$pf$click.points)
}
\arguments{
\item{ratio}{a numeric value, define the ratio of zomm in or out}

\item{resize}{Should the plot be resized?}

\item{p.center}{Click location centre}
}
\description{
Zoom in/out
}
\details{
if ratio < 1 then zoom in, if ratio > 1 then zoom out, if ratio = 1 then shift the plot.
}
\author{
Jason
}
