% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ClickOnZoom}
\alias{ClickOnZoom}
\title{Zoom in/out}
\usage{
ClickOnZoom(ratio = 1/2)
}
\arguments{
\item{ratio}{a numeric value, define the ratio of zomm in or out}
}
\description{
Zoom in/out when click the plot
}
\details{
if ratio < 1 then zoom in, if ratio > 1 then zoom out, if ratio = 1 then shift the plot.
}
\examples{
data("nzquakes")
iNZightPlot(Longitude,Latitude,data = nzquakes,colby = Depth, plottype = 'map',plot.features = list(maptype = 'roadmap'))
ClickOnZoom(ratio = 1)
}
\author{
Jason
}

