% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{col.fun}
\alias{col.fun}
\title{Color Specification}
\usage{
col.fun(data, color.index, display = "hue", na.fill = "white", offset = 0,
  col = "red")
}
\arguments{
\item{data}{a numeric vector that should lie on the range of [0,1]}

\item{color.index}{an integer vector from inzshapemap object}

\item{display}{a character value, the method for display colors. It should be choosen by one of the following display method: "hcl","hue","heat","rainbow","terrain","topo","cm","gray","r","n"}

\item{na.fill}{a character value that sepecify the color that use to fill the unmatch region/country.}

\item{offset}{a numeric value within the range of [0,1]}

\item{col}{the color for fill the match region/country, it only needs to be specify if display = 'hue'}
}
\value{
an color vector
}
\description{
Choose a type of color that helps read the map nicely.
}
\details{
hcl,HCL Color Specification, whith c = 35 and l = 85 see \link{hcl}. hue, when display = 'hue', then the 'col' arg need to be specified. The alpha will depend on the data, see \link{rgb}. rainbow,terrain,topo,cm are the method from \link{RColorBrewer}. r,n , the color filled randomly expect n will fill the entire map even the region is unmatch.
}
\author{
Jason Wen
}

