% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{color.bind}
\alias{color.bind}
\title{Plot an iNZight Map}
\usage{
color.bind(color, obj)
}
\arguments{
\item{obj}{object passed from iNZightPlot}

\item{gen}{other options passed from iNZightPlot}
}
\description{
draw a map by passing an iNZightPlot object
}
\details{
the function will also returns a global object which called global.objects
}
\author{
Jason Wen
}

