% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzmap.R
\name{create.inz.mapplot}
\alias{create.inz.mapplot}
\title{Create a map object to plot}
\usage{
create.inz.mapplot(obj)
}
\arguments{
\item{obj}{an object from within iNZightPlot}
}
\value{
Object
}
\author{
Tom Elliott
}

