% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{data.trans}
\alias{data.trans}
\title{data transformation}
\usage{
data.trans(x, transform = "linear")
}
\arguments{
\item{x}{a numeric value or vector.}

\item{transform}{the method for transformation, can be linear,log,sqrt,exp,power and normal.}
}
\value{
a transformed numeric vector.
}
\description{
Transform the data into the range of [0,1].
}
\author{
Jason Wen
}

