% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{findBestMatch}
\alias{findBestMatch}
\title{Find the pair of variables from the map and data that have the
highest number of matches. This function returns in form: (data
variable, map variable)}
\usage{
findBestMatch(data, map.data)
}
\description{
Find the pair of variables from the map and data that have the
highest number of matches. This function returns in form: (data
variable, map variable)
}
