% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{getNewMap}
\alias{getNewMap}
\title{Get an new map object}
\usage{
getNewMap(lat.lim, lon.lim, SCALE, type = c("roadmap", "mobile", "satellite",
  "terrain", "hybrid", "mapmaker-roadmap", "mapmaker-hybrid")[1], zoom)
}
\arguments{
\item{lat.lim}{the range of longitude}

\item{lon.lim}{the range of latitude}

\item{SCALE}{variable from \link{GetMap}, use the API's scale parameter to return higher-resolution map images. The scale value is multiplied with the size to determine the actual output size of the image in pixels, without changing the coverage area of the map}

\item{zoom}{variable from \link{GetMap}, Google maps zoom level.}

\item{variable}{from \link{GetMap}, type defines the type of map to construct. There are several possible maptype values, including satellite, terrain, hybrid, and mobile.}
}
\value{
a map object that assign as 'golbal.object'
}
\description{
Get a new map and save it into Global Env
}
\details{
a wrap function of \link{GetMap} from package 'RgoogleMaps'
}
\examples{
golbal.objects = list()
r.bbox = runif(4,-90,90)
getNewMap(r.bbox[1:2],r.bbox[3:4],2,zoom = 3)
}
\author{
Jason Wen
}

