% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{innerLim}
\alias{innerLim}
\title{inner limit}
\usage{
innerLim(obj, d.region)
}
\arguments{
\item{obj}{an inzshapemap object}

\item{d.region}{a character vector that sepecify the region or country.}
}
\value{
a numeric vector with length 4, the inner limit/bbox of the map.
}
\description{
Calculate the inner limit
}
\details{
the limit is computed by calculate the limit of the given region
}
\author{
Jason Wen
}

