% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapobject.R
\name{iNZightMap}
\alias{iNZightMap}
\title{Create an iNZight Map Object}
\usage{
iNZightMap(lat, lon, data, name = deparse(substitute(data)), ...)
}
\arguments{
\item{lat}{a formula specifying the name oe the latitude variable in \code{data}}

\item{lon}{a formula specifying the name oe the longitude variable in \code{data}}

\item{data}{a data frame}

\item{...}{exta arguments, not used}
}
\value{
data frame object with class \code{inzightmap}
}
\description{
Create an iNZight Map Object and pass to the plot
}
\details{
None.
}
\author{
tell029
}

