% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzightmap.R
\name{iNZightMapPlot}
\alias{iNZightMapPlot}
\alias{iNZightMapPlotRegion}
\alias{iNZightMapPlotPoint}
\title{Constructing an iNZightMapPlot object}
\usage{
iNZightMapPlot(data, map, type, ...)

iNZightMapPlotRegion(.data, .map, by.data, by.map)

iNZightMapPlotPoint(.data, .map, coord = c("lon", "lat"), crs = 4326)
}
\arguments{
\item{data}{Data values for each coordinate or region.}

\item{map}{A map to use with the values in \code{data}.}

\item{type}{Either \code{"region"} or \code{"point"} depending on the data.}

\item{...}{Any further required arguments for the type being constructed.}
}
\value{
An iNZightMapPlot object.
}
\description{
Constructing an iNZightMapPlot object
}
\section{Functions}{
\itemize{
\item \code{iNZightMapPlotRegion}: Constructs a iNZightMapPlot using region values.

\item \code{iNZightMapPlotPoint}: Constructs a iNZightMapPlot using point values.
}}

\examples{
## Region Data
library(gapminder)
data(gapminder)
map.world <- sf::st_as_sf(maps::map("world", ".", plot = FALSE, fill = TRUE))

world.mapplot <- iNZightMapPlot(gapminder, map.world, type = "region", 
                                by.data = "country", by.map = "ID")

plot(world.mapplot)

## Point Data
data(nzquakes)
map.nz <- sf::st_as_sf(maps::map("nz", ".", plot = FALSE, fill = TRUE))

nzquake.mapplot <- iNZightMapPlot(nzquakes, map.nz, type = "point", 
                                  coord = c("Longitude", "Latitude"))

plot(nzquake.mapplot)
}
