% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMapGG.R
\name{iNZightMapPlot}
\alias{iNZightMapPlot}
\alias{iNZightMapPlotRegion}
\title{Create iNZightMapPlot object}
\usage{
iNZightMapPlot(data, map, type, ...)

iNZightMapPlotRegion(data, map, by.data, by.map, simplification.level = 0.01,
  multiple.obs = FALSE, sequence.var = NULL, agg.type = "last")
}
\arguments{
\item{data}{Dataset with values for rows of the map object}

\item{map}{An sf object containing a row for each feature of the map}

\item{type}{What type of map is being passed in. Only \code{"region"} is implemented right now.}

\item{...}{Extra arguments required for the type of map being created.}

\item{by.data}{Variable name in the dataset that will be matched to \code{by.map} in the map}

\item{by.map}{Variable name in the map that will be matched to \code{by.data} in the dataset}

\item{simplification.level}{How much should the map be simplified by?}

\item{multiple.obs}{Does the dataset have multiple observations for each region of the map (i.e. observations from multiple years)}

\item{sequence.var}{If \code{multiple.obs = TRUE}, which variable identifies the different observations for a region?}

\item{agg.type}{If \code{multiple.obs = TRUE}, which aggregation should be used to produce one observation for each region.}
}
\description{
Create iNZightMapPlot object
}
\section{Functions}{
\itemize{
\item \code{iNZightMapPlotRegion}: Constructs a iNZightMapPlot using region values.
}}

