% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapemapobject.R
\name{iNZightShapeMap}
\alias{iNZightShapeMap}
\title{Create an iNZight Shape Map Object}
\usage{
iNZightShapeMap(location, shp.region, data.region, data)
}
\arguments{
\item{location}{the shape file}

\item{shp.region}{a character value, the column name in the region/country column of the shp file}

\item{data.region}{a character value, the column name in the region/country column of the data set}

\item{data}{the data set}
}
\value{
an iNZight Shape Map Object
}
\description{
Create an iNZightShapeMap object
}
\details{
details ....
}
\author{
Tom Elliott
}
