% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{latlon.xy}
\alias{latlon.xy}
\title{latlon.xy}
\usage{
latlon.xy(lat, lon, zoom)
}
\arguments{
\item{lat}{latitude values to convert}

\item{lon}{longitude values to convert}

\item{zoom}{the level of zoom used in the map}
}
\value{
a dataframe of x and y values
}
\description{
Transform the latitude and longitude into xy
}
\details{
Wrapper function for \link{LatLon2XY} from 'ggmap' package
}
\author{
Tom Elliott
}
