% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{latlon.xy}
\alias{latlon.xy}
\title{latlon.xy}
\usage{
latlon.xy(latlon, map)
}
\arguments{
\item{latlon}{the data set that use for plotting, the first cloumn needs to be latitude and the second column needs to be longitude.}

\item{map}{a map object from RgoogleMaps package.}
}
\value{
a list that contain the tramsformed x and y.
}
\description{
Transform the latitude and longitude into xy
}
\details{
wrap function for \link{LatLon2XY} from 'RgoogleMaps' package
}
\examples{
golbal.objects = list()
r.latlon = cbind(runif(100,-90,90),runif(100,-90,90))
r.bbox = runif(4,-90,90)
getNewMap(r.bbox[1:2],r.bbox[3:4],2,zoom = 3)
latlon.xy(r.latlon,global.objects$maps$map)
}
\author{
Jason Wen
}
