% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{lon.rescale}
\alias{lon.rescale}
\title{transform the latitude}
\usage{
lon.rescale(lon)
}
\arguments{
\item{lon}{a numeric value or a numeric vector of longitude}
}
\value{
a new longitude value or vector that used for plotting in map.
}
\description{
transform latitude inorder to get the biggest map as possibile
}
\details{
transform the longitude
}
\examples{
lon = runif(100,-360,360)
lon.rescale(lon)
}
\author{
Jason Wen
}

