% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{map.xylim}
\alias{map.xylim}
\title{get the limit of x-axis and y-axis}
\usage{
map.xylim(latR, lonR, SCALE)
}
\value{
return a list that contain the limit of x-axis and y-axis
}
\description{
return the limit of x-axis and y-axis of the plot.
}
\details{
the map.xylim function does not contain any arguments, and it returns the limit of x-axis and y-axis after the plot called.
}
\author{
Jason Wen
}

