% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{map.xylim}
\alias{map.xylim}
\title{get the limit of x-axis and y-axis}
\usage{
map.xylim(latR, lonR, SCALE)
}
\arguments{
\item{latR}{a numeric vector of length 2}

\item{lonR}{a numeric vector of length 2}

\item{SCALE}{a numeric value}
}
\value{
return a numeric of length 4 that contain the limit of x-axis and y-axis
}
\description{
get the limit of x-axis and y-axis
}
\note{
the limit is not the same as latitude/longitude, it is the limit of the plot window
}
\author{
Jason Wen
}
