% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{name.match}
\alias{name.match}
\title{Calaudate the bbox of a country}
\usage{
name.match(shp.region, data.region)
}
\arguments{
\item{obj}{the iNZight Shape Map Object}

\item{name}{the name of the country}
}
\value{
a 2*2 numeric matrix
}
\description{
Calaudate the bbox of a country
}
\examples{
los = 'C:/Users/yeamin/Documents/GitHub/iNZightMaps/data/world.rds'
shp = readRDS(los)
shp.region = shp$region
lod = 'C:/Users/yeamin/Desktop/Gapminder-2008.csv'
data = read.csv(lod,skip = 1)
name.match(shp.region,data$Country)
}
\author{
Jason
}
