% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{needNewMap}
\alias{needNewMap}
\title{title}
\usage{
needNewMap(bbox, window, size, SCALE, type)
}
\arguments{
\item{bbox}{}

\item{window}{}

\item{size}{}

\item{SCALE}{}

\item{type}{}
}
\value{
something
}
\description{
desc
}
\details{
det
}
\author{
tell029
}

