% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{order.match}
\alias{order.match}
\title{re-order the region name in shape file}
\usage{
order.match(shp.region, data.region)
}
\arguments{
\item{shp.region}{a character vector.}

\item{data.region}{a character vector.}
}
\value{
an integer vector.
}
\description{
re-order the region name in shape file
}
\author{
Jason Wen
}
