% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMapGG.R
\name{plot.iNZightMapPlot}
\alias{plot.iNZightMapPlot}
\title{Plot an iNZightMapPlot}
\usage{
\method{plot}{iNZightMapPlot}(x, colour.var = NULL, size.var = NULL,
  alpha.var = NULL, size.const = 1, alpha.const = 1,
  multiple.vars = FALSE, main = NULL, xlab = "Longitude",
  ylab = "Latitude", axis.labels = TRUE, datum.lines = TRUE,
  darkTheme = NULL, projection = "Default", palette = NULL,
  aggregate = FALSE, current.seq = NULL, sparkline.type = "Absolute",
  scale.limits = NULL, regions.to.plot = NULL,
  keep.other.regions = TRUE, label.var = NULL, scale.label = 1,
  scale.axis = 1, per.n = 1000, ...)
}
\arguments{
\item{x}{iNZightMapPlot object}

\item{colour.var}{Variable to colour the regions or points by}

\item{size.var}{If plotting a map of points, a variable to scale the points by}

\item{alpha.var}{If plotting a map of points, a variable to opacify the points by}

\item{size.const}{Size of plotted points (ignored if plotting regions or \code{size.var} is also passed)}

\item{alpha.const}{Alpha value of the underlying region map when plotting points}

\item{multiple.vars}{Are multiple variables being plotted?}

\item{main}{Main title for the plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{axis.labels}{Should the x- and y-axis labels be plotted}

\item{datum.lines}{Should the datum lines be plotte}

\item{darkTheme}{Enable dark background}

\item{projection}{Either \code{"Default"} or a proj4string}

\item{palette}{Palette to use}

\item{aggregate}{Is the plot an aggregate}

\item{current.seq}{Current value of the sequence variable or aggregation}

\item{sparkline.type}{Either \code{"Absolute"} or \code{"Relative"}}

\item{scale.limits}{Limits for the legend scale}

\item{regions.to.plot}{Which regions should be plotted?}

\item{keep.other.regions}{If \code{regions.to.plot} is not NULL, should regions that are neighbouring these regions still be plotted?}

\item{label.var}{Variable to label regions by}

\item{scale.label}{Scaling factor for region labels}

\item{scale.axis}{Scaling factor for title, axis labels, legend, etc.}

\item{per.n}{For dot density maps: how many people should each dot represent?}

\item{...}{additional arguments (ignored)}
}
\description{
Plot an iNZightMapPlot
}
