% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMapGG.R
\name{plot.iNZightMapPlot}
\alias{plot.iNZightMapPlot}
\title{Plot an iNZightMapPlot object}
\usage{
\method{plot}{iNZightMapPlot}(obj, colour.var = NULL, size.var = NULL,
  alpha.var = NULL, fill.const = NULL, colour.const = NULL,
  size.const = NULL, alpha.const = NULL, facet = NULL,
  multiple.vars = FALSE, main = NULL, xlab = "Longitude",
  ylab = "Latitude", axis.labels = TRUE, datum.lines = TRUE,
  darkTheme = NULL, projection = NULL, palette = NULL, label.title = "",
  aggregate = FALSE, current.seq = NULL)
}
\arguments{
\item{obj}{iNZightMapPlot object}

\item{colour.var}{String containing the variable name to map to
colour for either the regions or points (depending on the map
type)}

\item{size.var}{String containing the variable name to map to
point size}

\item{multiple.vars}{Is the plot of multiple variables?}

\item{main}{Title of the plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{axis.labels}{Should the axis labels be visible?}

\item{datum.lines}{Should the datum lines (grid behind the map) be
visible?}

\item{projection}{Either a PROJ4 string (or EPSG code?)}

\item{aggregate}{Is the plot some form of aggregate of the
original data (either mean/median/etc or a single year's data)}

\item{theme}{Either NULL for no theme or a string corresponding to
a theme. See [...]}
}
\value{
A ggplot object that can be plotted
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Daniel Barnett
}
