% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMapGG.R
\name{plot.iNZightMapPlot}
\alias{plot.iNZightMapPlot}
\title{Plot an iNZightMapPlot}
\usage{
\method{plot}{iNZightMapPlot}(obj, colour.var = NULL, size.var = NULL,
  alpha.var = NULL, fill.const = NULL, colour.const = NULL,
  size.const = 1, alpha.const = 1, facet = NULL, multiple.vars = FALSE,
  main = NULL, xlab = "Longitude", ylab = "Latitude",
  axis.labels = TRUE, datum.lines = TRUE, darkTheme = NULL,
  projection = "Default", palette = NULL, label.title = "",
  aggregate = FALSE, current.seq = NULL, sparkline.type = "Absolute",
  scale.limits = NULL, ci.plot = FALSE, regions.to.plot = NULL,
  keep.other.regions = TRUE, label.var = NULL)
}
\arguments{
\item{obj}{iNZightMapPlot object}

\item{colour.var}{Variable to colour the regions or points by}

\item{size.var}{If plotting a map of points, a variable to scale the points by}

\item{size.const}{Size of plotted points (ignored if plotting regions or \code{size.var} is also passed)}

\item{alpha.const}{Alpha value of the underlying region map when plotting points}

\item{multiple.vars}{Are multiple variables being plotted?}

\item{main}{Main title for the plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{axis.labels}{Should the x- and y-axis labels be plotted}

\item{datum.lines}{Should the datum lines be plotte}

\item{darkTheme}{Enable dark background}

\item{projection}{Either \code{"Default"} or a proj4string}

\item{palette}{Palette to use}

\item{aggregate}{Is the plot an aggregate}

\item{current.seq}{Current value of the sequence variable or aggregation}

\item{sparkline.type}{Either \code{"Absolute"} or \code{"Relative"}}

\item{scale.limits}{Limits for the legend scale}
}
\description{
Plot an iNZightMapPlot
}
