% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapemapobject.R
\name{plot.inzightshapemap}
\alias{plot.inzightshapemap}
\title{Create an iNZight Shape Map Object}
\usage{
\method{plot}{inzightshapemap}(x, variable, region, data, col.fun = "hue",
  transform = "linear", col.offset = 0.2, col = "red",
  na.fill = "white", full.map = TRUE, extend.ratio = 1, ...)
}
\arguments{
\item{x}{the iNZight Shape Map Object}

\item{region}{a variable or the column name of the region column}

\item{data}{the data set}

\item{col.fun}{a character value}

\item{full.map}{logical value.}

\item{variablie}{the variable or the column name in the data}
}
\description{
Create an iNZightShapeMap object
}
\details{
details ....
}
\author{
Tom Elliott
}

