% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzshapemap.R
\name{plot.inzshapemap}
\alias{plot.inzshapemap}
\title{Plot an iNZight Shape Map}
\usage{
\method{plot}{inzshapemap}(obj, gen)
}
\arguments{
\item{obj}{object passed from iNZightPlot}

\item{gen}{other options passed from iNZightPlot}
}
\description{
draw a shape file ...
}
\details{
details...
}
\author{
Jason Wen
}
