% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{region.bbox}
\alias{region.bbox}
\title{Calaudate the bbox of a country}
\usage{
region.bbox(obj, name)
}
\arguments{
\item{obj}{the iNZight Shape Map Object}

\item{name}{the name of the country}
}
\value{
a 2*2 numeric matrix
}
\description{
Calaudate the bbox of a country
}
\author{
Jason
}

