% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{sClickOnZoom}
\alias{sClickOnZoom}
\title{Zoom in/out}
\usage{
sClickOnZoom(ratio = 1/2, resize = FALSE)
}
\arguments{
\item{ratio}{a numeric value, define the ratio of zomm in or out}
}
\description{
Zoom in/out when click the plot
}
\details{
if ratio < 1 then zoom in, if ratio > 1 then zoom out, if ratio = 1 then shift the plot.
}
\author{
Jason
}
