% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{shape.extract}
\alias{shape.extract}
\title{extract and create a shape object}
\usage{
shape.extract(shp, column.index = 2)
}
\arguments{
\item{shp}{a SpatialPolygonsDataFrame object, see \link{readShapeSpatial}.}
}
\value{
a shape object.
}
\description{
extract the information from a SpatialPolygonsDataFrame object.
}
\details{
the function will also returns a global object which called global.objects.
}
\examples{
location = 'C:/Users/yeamin/Desktop/something/MED Final 20140226_region.shp'
shape.extract(location)
}
\author{
Jason Wen
}
