% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.fun.r
\name{subByLim}
\alias{subByLim}
\title{subset by limit}
\usage{
subByLim(obj, lim)
}
\arguments{
\item{obj}{an inzshapemap object.}

\item{lim}{a numeric vector of length 4.}
}
\value{
an inzshapemap object.
}
\description{
Subset by limit.
}
\author{
Jason Wen
}
